requirements.txt
streamlit
pandas
faiss-cpu
sentence-transformers
numpy
groq
